날씨 조회 api 변경 및 개선 작업 (#160)
* docs: 신규 날씨 예보 지역 정보 작성

* refactor: 위도, 경도 double 타입으로 변경

* feat: 신규 날씨 api 데이터 매핑을 위한 dto 구현

* feat: api 데이터 매퍼 클래스 구현

* refactor: 위도 경도 double 타입으로 변경

* feat: 신규 날씨 api 프로퍼티 정의

* feat: 신규 날씨 api 리퀘스터 구현

* refactor: 한국 위도 경도 경계값 설정 및 double 타입으로 변경

* refactor: 위도, 경도 double 타입 변경으로 인한 리팩토링

* feat: 신규 날씨 api 데이터 처리 서비스 레이어 구현

* refactor: 빈 이름 지정

* refactor: 빌더 위도, 경도 double 타입으로 수정

* refactor: 명시적으로 WeatherService 빈 주입하게 수정

* refactor: 위도, 경도 double 타입으로 변경

* feat: 날씨 조회 날짜 검증 부분 추가

* test: 날씨 api 변경에 따른 테스트 코드 수정

* feat: 기준 날짜 및 시간과 위치 정보로 날씨 조회하는 메서드 정의

* refactor: api key 프로퍼티 수정

* feat: 특정 도시에 해당되는 지역 조회하는 메서드 추가

* feat: 매시간의 15분 때마다 날씨 정보를 캐싱하는 로직 구현

* refactor: 시간 유효성 검증 주석처리 제거

* docs: 경기도 성남구 정보 추가

* docs: 경기도 성남시 중복되는 부분 제거

+++ src/main/java/com/example/fashionforecastbackend/weather/dto/response/WeatherForecast.java
@@ -14,8 +14,8 @@ public record WeatherForecast(
 	String pcp,
 	RainType rainType,
 	SkyStatus skyStatus,
-	int nx,
-	int ny
+	double nx,
+	double ny
 ) {
 	public static WeatherForecast from(Weather weather) {
 		return new WeatherForecast(
